[fix] GH-28571- Insert head slice below select(len() <cmp> n) to enable slice pushdown
